# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "TakeIn Resturant",
    "description": "A Food Delivering App"
  },
  "servers": [
    {
      "url": "http://takein-prod-public-alb-373215541.ap-southeast-1.elb.amazonaws.com/api/v1"
    },
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [{
    "bearerAuth": []
  }],
  "paths": {
    "/user/send/otp": {
      "post": {
        "tags": ["OTP & SIGNUP SCREEN"],
        "summary": "Sending OTP to Mobile Number",
        "requestBody": {
          "description": "OTP Send Payload",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/OTP_SEND"
              }
            }
          }
        },
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "responses": {
          "200": {
            "description": "OTP sent successfully!",
            "schema": {
              "$ref": "#/definitions/OTP_SEND"
            }
          },
          "400": {
            "description": "There is some error while sending the OTP."
          }
        }
      }
    },
    "/user/verify/otp": {
      "post": {
        "tags": ["OTP & SIGNUP SCREEN"],
        "summary": "Verifying OTP",
        "requestBody": {
          "description": "OTP Verify Payload",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/OTP_VERIFY"
              }
            }
          }
        },
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "responses": {
          "200": {
            "description": "OTP Verified successfully!",
            "schema": {
              "$ref": "#/definitions/OTP_VERIFY"
            }
          },
          "400": {
            "description": "There is some error while verifying the OTP."
          }
        }
      }
    },
    "/user/resend/otp": {
      "post": {
        "tags": ["OTP & SIGNUP SCREEN"],
        "summary": "Resending OTP",
        "requestBody": {
          "description": "OTP Resend Payload",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/OTP_RESEND"
              }
            }
          }
        },
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "responses": {
          "200": {
            "description": "OTP Resent successfully!",
            "schema": {
              "$ref": "#/definitions/OTP_RESEND"
            }
          },
          "400": {
            "description": "There is some error while sending the OTP."
          }
        }
      }
    },
    "/user/sign-up": {
      "post": {
        "tags": ["OTP & SIGNUP SCREEN"],
        "summary": "User Sign Up",
        "requestBody": {
          "description": "User Sign Up form",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/USER_SIGNUP"
              }
            }
          }
        },
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "responses": {
          "200": {
            "description": "You have successfully completed sign up!",
            "schema": {
              "$ref": "#/definitions/USER_SIGNUP"
            }
          },
          "400": {
            "description": "Please provide data for all the required field to proceed sign up."
          }
        }
      }
    },
...
    "/user/address": {
      "post": {
        "tags": ["COOKIEMONSTER_ADDRESS"],
        "summary": "Adding CookieMonster Address",
        "requestBody": {
          "description": "Address Payload",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/COOKIEMONSTER_ADDRESS"
              }
            }
          }
        },
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "responses": {
          "200": {
            "description": "You have successfully saved your Address!",
            "schema": {
              "$ref": "#/definitions/COOKIEMONSTER_ADDRESS"
            }
          },
          "400": {
            "description": "There is some error while saving Address."
          }
        }
      },
      "get": {
        "tags": ["COOKIEMONSTER_ADDRESS"],
        "summary": "Fetching Cookie Monster Address",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "responses": {
          "200": {
            "description": "You have successfully fetched Cookie Monster Address.",
            "schema": {
              "$ref": "#/definitions/COOKIEMONSTER_ADDRESS"
            }
          },
          "400": {
            "description": "Error in fetching Cookie Monster Address."
          }
        }
      }
    },
    "/user/current/address": {
      "post": {
        "tags": ["COOKIEMONSTER_ADDRESS"],
        "summary": "Adding CookieMonster Current Address",
        "requestBody": {
          "description": "Address Payload",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/COOKIEMONSTER_CURRENT_ADDRESS"
              }
            }
          }
        },
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "responses": {
          "200": {
            "description": "You have successfully saved your Current Address!",
            "schema": {
              "$ref": "#/definitions/COOKIEMONSTER_ADDRESS"
            }
          },
          "400": {
            "description": "There is some error while saving Address."
          }
        }
      }
    },
    "/user/address/{id}": {
      "put": {
        "summary": "Update Cookie Monster Address",
        "tags": ["COOKIEMONSTER_ADDRESS"],
        "requestBody": {
          "description": "Restaurant Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/COOKIEMONSTER_ADDRESS"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Address Id",
            "schema": {
              "$ref": "#/definitions/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated cookie Monster Address.",
            "schema": {
              "$ref": "#/definitions/COOKIEMONSTER_ADDRESS"
            }
          },
          "400": {
            "description": "Failed to update cookie montser address"
          },
          "404": {
            "description": "Failed. Addess not found."
          }
        }
      },
      "delete": {
        "summary": "Delete Address of Cookie Monster",
        "tags": ["COOKIEMONSTER_ADDRESS"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Cookie Monster Address Id.",
            "schema": {
              "$ref": "#/definitions/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully Deleted Cookie Monster Address Id.",
            "schema": {
              "$ref": "#/definitions/COOKIEMONSTER_ADDRESS"
            }
          },
          "400": {
            "description": "Failed to Delete Cookie Monster Id"
          },
          "404": {
            "description": "Failed. Address Not found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "OTP_SEND": {
      "type": "object",
      "properties": {
        "payload": {
          "type": "object",
          "properties": {
            "mobileNumber": {
              "type": "number",
              "example": 9033528947
            },
            "countryCode": {
              "type": "string",
              "example": "+91"
            }
          }
        }
      }
    },
    "OTP_VERIFY": {
      "type": "object",
      "properties": {
        "payload": {
          "type": "object",
          "properties": {
            "mobileNumber": {
              "type": "number",
              "example": 9033528947
            },
            "otp": {
              "type": "number",
              "example": "4785"
            },
            "userId": {
              "type": "string",
              "example": "610a7d8f0a4d221c40e7d849"
            }
          }
        }
      }
    },
    "OTP_RESEND": {
      "type": "object",
      "properties": {
        "payload": {
          "type": "object",
          "properties": {
            "mobileNumber": {
              "type": "number",
              "example": 9033528947
            },
            "userId": {
              "type": "string",
              "example": "610a7d8f0a4d221c40e7d849"
            }
          }
        }
      }
    },
...
    "USER_SIGNUP": {
      "type": "object",
      "properties": {
        "payload": {
          "type": "object",
          "properties": {
            "firstName": {
              "type": "string",
              "example": "Test"
            },
            "lastName": {
              "type": "string",
              "example": "Test"
            },
            "email": {
              "type": "string",
              "example": "test@gmail.com"
            },
            "referralCode": {
              "type": "string",
              "example": "AFY25Q"
            },
            "userId": {
              "type": "string",
              "example": "610a7d8f0a4d221c40e7d849"
            }
          }
        }
      }
    },
...
    "id": {
      "properties": {
        "uuid": {
          "type": "string"
        }
      }
    },
    "COOKIEMONSTER_ADDRESS": {
      "type": "object",
      "properties": {
        "payload": {
          "type": "object",
          "properties": {
            "address": {
              "type": "object",
              "required": [
                "nickName",
                "addressLine1",
                "addressLine2",
                "phoneNumber"
              ],
              "properties": {
                "nickName": {
                  "type": "string",
                  "example": "Home"
                },
                "addressLine1": {
                  "type": "string",
                  "example": "Burj Khalifa"
                },
                "addressLine2": {
                  "type": "string",
                  "example": "Dubai"
                },
                "phoneNumber": {
                  "type": "number",
                  "example": 9033528974
                },
                "location": {
                  "type": "object",
                  "properties": {
                    "coordinates": {
                      "type": "double",
                      "example": [23.8426943813556, 61.4644750214197]
                    }
                  }
                },
                "isDefault": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        }
      }
    },
    "COOKIEMONSTER_CURRENT_ADDRESS": {
      "type": "object",
      "properties": {
        "location": {
          "type": "object",
          "properties": {
            "coordinates": {
              "type": "double",
              "example": [23.8426943813556, 61.4644750214197]
            }
          }
        }
      }
    },
...
  }
}